refactor: Checkbox, Range bar -> 커스텀 훅으로 분리
